# src/TaskFlow.Server/Contracts/ReferResultDto.cs
namespace TaskFlow.Server;

public sealed class ReferResultDto
{
    public string InstanceId { get; set; } = "";
    public string DefinitionKey { get; set; } = "";
    public TaskDto? Task { get; set; }
    public List<TaskDto> Tasks { get; set; } = [];
}