# src/TaskFlow.Server/Contracts/UserProcessListDto.cs
namespace TaskFlow.Server;

public sealed class UserProcessListDto
{
    public string User { get; set; } = "";
    public string State { get; set; } = "";
    public int Open { get; set; }
    public int Closed { get; set; }
    public int NotStarted { get; set; }
    public int Total { get; set; }
    public List<ProcessInstanceDetailDto> Instances { get; set; } = [];
}